refactor: add cli/standalone folder
